--- src/main/java/util/WinningLottoGenerator.java
@@ -1,19 +1,0 @@
-package util;
-
-import constant.ConstantNumber;
-import domain.Lotto;
-
-import java.util.List;
-
-import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
-
-public class WinningLottoGenerator {
-    public Lotto create() {
-        List<Integer> winningNumbers = pickUniqueNumbersInRange(
-                ConstantNumber.MINIMUM_WINNING_NUMBER.get(),
-                ConstantNumber.MAXIMUM_WINNING_NUMBER.get(),
-                ConstantNumber.WINNING_NUMBER_COUNT.get()
-                );
-        return new Lotto(winningNumbers);
-    }
-}
